# [Discover] Improve data grid render performance (`flattenHit` and `formatHit`)