--- a/packages/kbn-data-service/index.ts
+++ b/packages/kbn-data-service/index.ts
@@ -10,4 +10,9 @@
 export type { GetConfigFn } from './src/types';
 export { UI_SETTINGS } from './src/constants';
 export { getEsQueryConfig } from './src/es_query';
-export { tabifyDocs, flattenHit } from './src/search/tabify';
+export {
+  tabifyDocs,
+  flattenHit,
+  getFlattenedFieldsComparator,
+  type FlattenedFieldsComparator,
+} from './src/search/tabify';