--- a/packages/kbn-data-service/src/search/tabify/tabify_docs.ts
+++ b/packages/kbn-data-service/src/search/tabify/tabify_docs.ts
@@ -98,7 +98,11 @@ function flattenAccum(
  * @param indexPattern The index pattern for the requested index if available.
  * @param params Parameters how to flatten the hit
  */
-export function flattenHit(hit: Hit, indexPattern?: DataView, params?: TabifyDocsOptions) {
+export function flattenHit(
+  hit: Hit,
+  indexPattern?: DataView,
+  params?: TabifyDocsOptions & { flattenedFieldsComparator?: FlattenedFieldsComparator }
+) {
   const flat = {} as Record<string, any>;
 
   flattenAccum(flat, hit.fields || {}, '', indexPattern, params);
@@ -147,49 +151,55 @@ export function flattenHit(hit: Hit, indexPattern?: DataView, params?: TabifyDoc
 
   // Use a proxy to make sure that keys are always returned in a specific order,
   // so we have a guarantee on the flattened order of keys.
-  return makeProxy(flat, indexPattern);
+  return makeProxy(flat, indexPattern, params?.flattenedFieldsComparator);
 }
 
-function makeProxy(flat: Record<string, any>, indexPattern?: DataView) {
-  const getComparator = () => {
-    const metaFields = new Set(indexPattern?.metaFields);
-    const lowerMap = new Map<string, string>();
-    let aLower: string | undefined;
-    let bLower: string | undefined;
+export const getFlattenedFieldsComparator = (indexPattern?: DataView) => {
+  const metaFields = new Set(indexPattern?.metaFields);
+  const lowerMap = new Map<string, string>();
+  let aLower: string | undefined;
+  let bLower: string | undefined;
 
-    const compareLower = (a: string, b: string) => {
-      aLower = lowerMap.get(a);
-      if (aLower === undefined) {
-        aLower = a.toLowerCase();
-        lowerMap.set(a, aLower);
-      }
-      bLower = lowerMap.get(b);
-      if (bLower === undefined) {
-        bLower = b.toLowerCase();
-        lowerMap.set(b, bLower);
-      }
-      return aLower < bLower ? -1 : aLower > bLower ? 1 : 0;
-    };
-
-    return (a: string | symbol, b: string | symbol) => {
-      if (typeof a === 'symbol' || typeof b === 'symbol') {
-        return 0;
-      }
-      const aIsMeta = metaFields.has(a);
-      const bIsMeta = metaFields.has(b);
-      if (aIsMeta && bIsMeta) {
-        return compareLower(a, b);
-      }
-      if (aIsMeta) {
-        return 1;
-      }
-      if (bIsMeta) {
-        return -1;
-      }
-      return compareLower(a, b);
-    };
+  const compareLower = (a: string, b: string) => {
+    aLower = lowerMap.get(a);
+    if (aLower === undefined) {
+      aLower = a.toLowerCase();
+      lowerMap.set(a, aLower);
+    }
+    bLower = lowerMap.get(b);
+    if (bLower === undefined) {
+      bLower = b.toLowerCase();
+      lowerMap.set(b, bLower);
+    }
+    return aLower < bLower ? -1 : aLower > bLower ? 1 : 0;
   };
 
+  return (a: string | symbol, b: string | symbol) => {
+    if (typeof a === 'symbol' || typeof b === 'symbol') {
+      return 0;
+    }
+    const aIsMeta = metaFields.has(a);
+    const bIsMeta = metaFields.has(b);
+    if (aIsMeta && bIsMeta) {
+      return compareLower(a, b);
+    }
+    if (aIsMeta) {
+      return 1;
+    }
+    if (bIsMeta) {
+      return -1;
+    }
+    return compareLower(a, b);
+  };
+};
+
+export type FlattenedFieldsComparator = ReturnType<typeof getFlattenedFieldsComparator>;
+
+function makeProxy(
+  flat: Record<string, any>,
+  indexPattern?: DataView,
+  flattenedFieldsComparator?: FlattenedFieldsComparator
+) {
   let cachedKeys: Array<string | symbol> | undefined;
 
   return new Proxy(flat, {
@@ -203,7 +213,9 @@ function makeProxy(flat: Record<string, any>, indexPattern?: DataView) {
     },
     ownKeys: (target) => {
       if (!cachedKeys) {
-        cachedKeys = Reflect.ownKeys(target).sort(getComparator());
+        cachedKeys = Reflect.ownKeys(target).sort(
+          flattenedFieldsComparator ?? getFlattenedFieldsComparator(indexPattern)
+        );
       }
       return cachedKeys;
     },