--- a/packages/kbn-data-service/src/search/tabify/tabify_docs.ts
+++ b/packages/kbn-data-service/src/search/tabify/tabify_docs.ts
@@ -171,19 +171,15 @@ function makeProxy(flat: Record<string, any>, indexPattern?: DataView) {
   return new Proxy(flat, {
     defineProperty: (...args) => {
       cachedKeys = undefined;
-      console.log('defineProperty');
       return Reflect.defineProperty(...args);
     },
     deleteProperty: (...args) => {
       cachedKeys = undefined;
-      console.log('deleteProperty');
       return Reflect.deleteProperty(...args);
     },
     ownKeys: (target) => {
       if (!cachedKeys) {
         cachedKeys = Reflect.ownKeys(target).sort(comparator);
-      } else {
-        console.log('cached');
       }
       return cachedKeys;
     },