--- a/packages/kbn-discover-utils/src/utils/build_data_record.ts
+++ b/packages/kbn-discover-utils/src/utils/build_data_record.ts
@@ -8,7 +8,11 @@
  */
 
 import type { DataView } from '@kbn/data-views-plugin/common';
-import { flattenHit } from '@kbn/data-service';
+import {
+  type FlattenedFieldsComparator,
+  flattenHit,
+  getFlattenedFieldsComparator,
+} from '@kbn/data-service';
 import type { DataTableRecord, EsHitRecord } from '../types';
 import { getDocId } from './get_doc_id';
 
@@ -21,12 +25,16 @@ import { getDocId } from './get_doc_id';
 export function buildDataTableRecord(
   doc: EsHitRecord,
   dataView?: DataView,
-  isAnchor?: boolean
+  isAnchor?: boolean,
+  options?: { flattenedFieldsComparator?: FlattenedFieldsComparator }
 ): DataTableRecord {
   return {
     id: getDocId(doc),
     raw: doc,
-    flattened: flattenHit(doc, dataView, { includeIgnoredValues: true }),
+    flattened: flattenHit(doc, dataView, {
+      includeIgnoredValues: true,
+      flattenedFieldsComparator: options?.flattenedFieldsComparator,
+    }),
     isAnchor,
   };
 }
@@ -45,8 +53,9 @@ export function buildDataTableRecordList<T extends DataTableRecord = DataTableRe
   dataView?: DataView;
   processRecord?: (record: DataTableRecord) => T;
 }): DataTableRecord[] {
+  const buildRecordOption = { flattenedFieldsComparator: getFlattenedFieldsComparator(dataView) };
   return records.map((doc) => {
-    const record = buildDataTableRecord(doc, dataView);
+    const record = buildDataTableRecord(doc, dataView, undefined, buildRecordOption);
     return processRecord ? processRecord(record) : record;
   });
 }