--- a/packages/kbn-discover-utils/src/utils/format_hit.ts
+++ b/packages/kbn-discover-utils/src/utils/format_hit.ts
@@ -18,6 +18,15 @@ import type {
 } from '../types';
 import { formatFieldValue } from './format_value';
 
+// We use a special type here allowing formattedValue to be undefined because
+// we want to avoid formatting values which will not be shown to users since
+// it can be costly, and instead only format the ones which will be rendered
+type PartialHitPair = [
+  fieldDisplayName: string,
+  formattedValue: string | undefined,
+  fieldName: string | null
+];
+
 const formattedHitCache = new WeakMap<
   EsHitRecord,
   { formattedHit: FormattedHit; maxEntries: number }
@@ -48,72 +57,69 @@ export function formatHit(
 
   const highlights = hit.raw.highlight ?? {};
   const flattened = hit.flattened;
-  const flattenedKeys = Object.keys(flattened);
-  const formattedHit: FormattedHit = [];
+  const renderedPairs: PartialHitPair[] = [];
+  const otherPairs: PartialHitPair[] = [];
 
-  let totalCount = 0;
-
-  const maybeAddToFormattedHit = (key: string) => {
+  // Add each flattened field into the corresponding array for rendered or other pairs,
+  // depending on whether the original hit had a highlight for it. That way we can ensure
+  // highlighted fields are shown first in the document summary.
+  for (const key of Object.keys(flattened)) {
+    // Retrieve the (display) name of the fields, if it's a mapped field on the data view
     const field = dataView.fields.getByName(key);
+    const displayKey = field?.displayName;
+    const pairs = highlights[key] ? renderedPairs : otherPairs;
 
-    // If the field was a mapped field, we validate it against the fieldsToShow list, if not
-    // we always include it into the result.
-    if (field?.displayName) {
+    // If the field is a mapped field, we first check if it should be shown,
+    // if not we always include it into the result.
+    if (displayKey) {
       if (shouldShowFieldHandler(key)) {
-        if (++totalCount > maxEntries) {
-          return;
-        }
-
-        formattedHit.push([
-          // Retrieve the (display) name of the fields, if it's a mapped field on the data view
-          field.displayName,
-          // Format the raw value using the regular field formatters for that field
-          formatFieldValue(flattened[key], hit.raw, fieldFormats, dataView, field),
-          key,
-        ]);
+        pairs.push([displayKey, undefined, key]);
       }
     } else {
-      if (++totalCount > maxEntries) {
-        return;
-      }
-
-      formattedHit.push([
-        key,
-        // Format the raw value using the regular field formatters for that field
-        formatFieldValue(flattened[key], hit.raw, fieldFormats, dataView, field),
-        key,
-      ]);
-    }
-  };
-
-  // Add highlighted fields first if any exist
-  if (Object.keys(highlights).length) {
-    for (const key of flattenedKeys) {
-      if (highlights[key]) {
-        maybeAddToFormattedHit(key);
-      }
+      pairs.push([key, undefined, key]);
     }
   }
 
-  // Add remaining fields after
-  for (const key of flattenedKeys) {
-    if (!highlights[key]) {
-      maybeAddToFormattedHit(key);
+  const totalLength = renderedPairs.length + otherPairs.length;
+
+  // Truncate the renderedPairs if it exceeds the maxEntries,
+  // otherwise fill it up with otherPairs until it reaches maxEntries
+  if (renderedPairs.length > maxEntries) {
+    renderedPairs.length = maxEntries;
+  } else if (renderedPairs.length < maxEntries && otherPairs.length) {
+    for (let i = 0; i < otherPairs.length && renderedPairs.length < maxEntries; i++) {
+      renderedPairs.push(otherPairs[i]);
     }
+  }
+
+  // Now format only the values which will be shown to the user
+  for (const pair of renderedPairs) {
+    const key = pair[2]!;
+
+    // Format the raw value using the regular field formatters for that field
+    pair[1] = formatFieldValue(
+      flattened[key],
+      hit.raw,
+      fieldFormats,
+      dataView,
+      dataView.getFieldByName(key)
+    );
   }
 
   // If document has more formatted fields than configured via MAX_DOC_FIELDS_DISPLAYED we cut
   // off additional fields and instead show a summary how many more field exists.
-  if (totalCount > maxEntries) {
-    formattedHit.push([
+  if (totalLength > maxEntries) {
+    renderedPairs.push([
       i18n.translate('discover.formatHit.moreFields', {
         defaultMessage: 'and {count} more {count, plural, one {field} other {fields}}',
-        values: { count: totalCount - maxEntries },
+        values: { count: totalLength - maxEntries },
       }),
       '',
       null,
     ]);
   }
+
+  const formattedHit = renderedPairs as FormattedHit;
 
   formattedHitCache.set(hit.raw, { formattedHit, maxEntries });