--- a/src/plugins/discover/public/embeddable/initialize_fetch.ts
+++ b/src/plugins/discover/public/embeddable/initialize_fetch.ts
@@ -11,12 +11,11 @@ import { BehaviorSubject, combineLatest, lastValueFrom, switchMap, tap } from 'r
 
 import { KibanaExecutionContext } from '@kbn/core/types';
 import {
-  buildDataTableRecord,
+  buildDataTableRecordList,
   SEARCH_EMBEDDABLE_TYPE,
   SEARCH_FIELDS_FROM_SOURCE,
   SORT_DEFAULT_ORDER_SETTING,
 } from '@kbn/discover-utils';
-import { EsHitRecord } from '@kbn/discover-utils/types';
 import { isOfAggregateQueryType, isOfQueryType } from '@kbn/es-query';
 import { i18n } from '@kbn/i18n';
 import { RequestAdapter } from '@kbn/inspector-plugin/common';
@@ -207,7 +206,12 @@ export function initializeFetch({
 
           return {
             warnings: interceptedWarnings,
-            rows: resp.hits.hits.map((hit) => buildDataTableRecord(hit as EsHitRecord, dataView)),
+            rows: buildDataTableRecordList({
+              records: resp.hits.hits,
+              dataView,
+              processRecord: (record) =>
+                discoverServices.profilesManager.resolveDocumentProfile({ record }),
+            }),
             hitCount: resp.hits.total as number,
             fetchContext,
           };